[refactor] moved code to correct locations
